requirements.txt
boto3==1.34.91
Flask==2.3.2
llama_index==0.10.31
python-dotenv==1.0.1
spacy==3.7.4
